--- src/common/helper/dtos/query-pagination.dto.ts
@@ -1,14 +1,0 @@
-import { Type } from 'class-transformer';
-import { IsOptional, IsNumber } from 'class-validator';
-
-export class QueryPaginateDto {
-  @IsOptional()
-  @IsNumber()
-  @Type(() => Number)
-  page?: number;
-
-  @IsOptional()
-  @IsNumber()
-  @Type(() => Number)
-  limit?: number;
-}
