+++ src/modules/requests/respositories/requests.repository.ts
@@ -1,7 +1,7 @@
 import { Repository } from 'typeorm';
 import { Request } from '../entities/request.entity';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Pagination } from '../../../common/helpers/pagination.helper';
+import { Pagination } from '../../../common/helpers/pagination';
 import { UserAccess } from '../../../common/interfaces/auth-user.interface';
 export class RequestsRepository extends Repository<Request> {
   constructor(
